requirements.txt
# ===== CORE BACKEND =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# ===== DATABASE (MongoDB) =====
motor==3.3.2
beanie==1.24.0
pymongo==4.6.1

# ===== RHYME GENERATION =====
pronouncing==0.2.0
nltk==3.8.1

# ===== LLM (Gemini) =====
google-generativeai==0.3.1

# ===== TTS (RECOMMENDED: Edge TTS - 100% FREE!) =====
edge-tts==6.1.9

# ===== AUDIO PROCESSING =====
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
numpy==1.26.2
scipy==1.11.4
pyrubberband==0.3.0

# ===== KARAOKE ALIGNMENT (CRITICAL) =====
aeneas==1.7.3

# ===== ASYNC TASKS (REQUIRED) =====
celery==5.3.4
redis==5.0.1

# ===== UTILITIES =====
requests==2.31.0
aiofiles==23.2.1
